requirements.txt
attrs==24.2.0
blinker==1.8.2
cachelib==0.9.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-iconify==0.1.2
dash-table==5.0.0
dash-tools==1.12.0
dash_mantine_components==0.15.0
distlib==0.3.9
dnspython==2.6.1
docopt==0.6.2
et-xmlfile==1.1.0
filelock==3.16.1
Flask==3.0.3
Flask-Caching==2.3.0
gunicorn==23.0.0
identify==2.6.3
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==2.1.1
openpyxl==3.1.5
packaging==21.3
pandas==2.2.3
pipreqs==0.4.12
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
pre-commit==2.16.0
py==1.11.0
pymongo==4.10.1
pyparsing==3.2.0
pytest==6.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
retrying==1.3.4
ruamel.yaml==0.17.21
setuptools==75.1.0
six==1.16.0
tenacity==9.0.0
termcolor==1.1.0
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.28.0
visdcc==0.0.50
Werkzeug==3.0.4
yarg==0.1.10
yolk3k==0.9
zipp==3.20.2
